requirements.txt
django==3.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0